refactor: toast 코드 정리
