+++ src/app/template.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { type PropsWithChildren } from 'react';
-import { ToastProvider } from '@/components/common/Toast/ToastProvider';
+import { ToastProvider } from '@/providers/ToastProvider';
 import { theme } from '@/styles';
 import styled from '@emotion/styled';
 import { OverlayProvider } from '@toss/use-overlay';
