+++ src/components/common/Toast/index.ts
@@ -1,1 +1,2 @@
 export * from './Toast';
+export * from './ToastContainer';
