+++ src/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './StyleProvider';
 export * from './SuspensiveProvider';
+export * from './ToastProvider';
